Extract render and controller validation helpers
