+++ CMakeLists.txt
@@ -204,8 +204,10 @@ target_sources(
         src/game_data_standard_options.c
         src/game_presentation.c
         src/game_data_validation_brush_common.c
+        src/game_data_validation_controller_components.c
         src/game_data_validation_brush_worlds.c
         src/game_data_validation_editor_metadata.c
+        src/game_data_validation_render_components.c
         src/game_data_validation_scene_editor.c
         src/game_data_validation.c
         src/gltf_loader.c
